+++ app-dev/js/context-info.js
@@ -1,12 +1,12 @@
 (() => {
 
-  const $MinorInfoBar = document.getElementById('minor-info-bar');
+  const $MinorInfoBar = document.getElementById('minor-info-bar_unc2741');
 
-  const $ContextMenu = document.getElementById('context-menu');
-  const $ContextMenuLinkOpen = $ContextMenu.querySelector('.a-view .open .inner-container');
-  const $ContextMenuImgOpen = $ContextMenu.querySelector('.img-view .open .inner-container');
+  const $ContextMenuLinkOpen = N.$ContextMenu.querySelector('.a-view .open a');
+  const $ContextMenuAnchorGoTo = N.$ContextMenu.getElementsByClassName('go-to-anchor')[0];
+  const $ContextMenuImgOpen = N.$ContextMenu.querySelector('.img-view .open a');
 
-  const $PreviewLarge = document.getElementById('preview-large');
+  const $PreviewLarge = document.getElementById('preview-large_unc2741');
   const $PreviewLargeImg = $PreviewLarge.getElementsByTagName('img')[0];
 
 
@@ -15,44 +15,57 @@
     let $Target = Parameter.$ElementBase;
 
     while (
-      ! $Target.classList.contains('tab')
-      && $Target.tagName !== 'A'
+      $Target.tagName !== 'A'
       && $Target.tagName !== 'IMG'
+      && ! N.RegExpHeading.test($Target.tagName)
       && ! $Target.classList.contains('emoji')
       && ! $Target.classList.contains('preview-list-item')
+      && ! $Target.classList.contains('tab')
+      && $Target !== N.$Workspace
       && $Target !== document.body
     ) {
       $Target = $Target.parentNode;
     }
 
     if (
-      $Target.classList.contains('tab')
-      || $Target.tagName === 'A'
+      $Target.tagName === 'A'
       || $Target.tagName === 'IMG'
+      || N.RegExpHeading.test($Target.tagName)
       || $Target.classList.contains('emoji')
       || $Target.classList.contains('preview-list-item')
+      || $Target.classList.contains('tab')
+      || $Target === N.$Workspace
     ) {
       let TargetInfo = {};
 
-      if ($Target.classList.contains('tab')) {
-        TargetInfo = {
-          $Target: $Target,
-          strTargetType: 'tab',
-          strTargetValue: N.arrDocs[$Target.dataset.id].strPath
-        };
-      }
-      else if ($Target.tagName === 'A') {
-        TargetInfo = {
-          $Target: $Target,
-          strTargetType: 'a',
-          strTargetValue: $Target.href
-        };
+      if ($Target.tagName === 'A') {
+        // http://stackoverflow.com/questions/3861076/how-can-i-extract-text-after-hash-in-the-href-part-from-a-tag/3861095#3861095
+        if ($Target.hash) {
+          TargetInfo = {
+            $Target: $Target,
+            strTargetType: 'anchor-link',
+            strTargetValue: $Target.hash
+          };
+        } else {
+          TargetInfo = {
+            $Target: $Target,
+            strTargetType: 'a',
+            strTargetValue: $Target.href
+          };
+        }
       }
       else if ($Target.tagName === 'IMG') {
         TargetInfo = {
           $Target: $Target,
           strTargetType: 'img',
           strTargetValue: $Target.src
+        };
+      }
+      else if (N.RegExpHeading.test($Target.tagName)) {
+        TargetInfo = {
+          $Target: $Target,
+          strTargetType: 'anchor',
+          strTargetValue: '#' + $Target.id
         };
       }
       else if ($Target.classList.contains('emoji')) {
@@ -67,6 +80,20 @@
           $Target: $Target,
           strTargetType: 'preview-list-item',
           strTargetValue: $Target.getElementsByTagName('img')[0].src
+        };
+      }
+      else if ($Target.classList.contains('tab')) {
+        TargetInfo = {
+          $Target: $Target,
+          strTargetType: 'tab',
+          strTargetValue: N.arrDocs[$Target.dataset.id].strPath
+        };
+      }
+      else if ($Target === N.$Workspace) {
+        TargetInfo = {
+          $Target: $Target,
+          strTargetType: 'text',
+          strTargetValue: null
         };
       }
 
@@ -95,8 +122,8 @@
 
 
   document.addEventListener('click', (Event) => {
-    if ( ! Event.target.classList.contains('inner-container')) {
-      $ContextMenu.classList.remove('active');
+    if ( ! Event.target.classList.contains('menu-list-item')) {
+      N.$ContextMenu.classList.remove('active');
     }
   });
 
@@ -107,11 +134,11 @@
 
     if (TargetInfo && TargetInfo.strTargetValue) {
 
-      if ( ! TargetInfo.$Target.classList.contains('no-info-bar') ) {
+      if (TargetInfo.$Target.classList.contains('no-info-bar') || TargetInfo.$Target.classList.contains('no-context-info')) {
+        $MinorInfoBar.classList.remove('active');
+      } else {
         $MinorInfoBar.textContent = TargetInfo.strTargetValue;
         $MinorInfoBar.classList.add('active');
-      } else {
-        $MinorInfoBar.classList.remove('active');
       }
 
       if (TargetInfo.strTargetType === 'preview-list-item') {
@@ -135,51 +162,93 @@
   document.addEventListener('contextmenu', (Event) => {
     const TargetInfo = funcFindRealTarget({ $ElementBase: Event.target });
 
-    if (TargetInfo && ! TargetInfo.$Target.classList.contains('no-context-menu') ) {
-      $ContextMenu.dataset.target = TargetInfo.strTargetType;
+    if (TargetInfo) {
 
-      // No need to check TargetInfo.strTargetValue, the user can't directly add an empty value and the document content is cleared at input.
-      if (TargetInfo.strTargetType === 'tab') {
-        N.$Header.classList.add('active');
+      if (TargetInfo.$Target.classList.contains('no-context-menu') || TargetInfo.$Target.classList.contains('no-context-info')) {
+        N.$ContextMenu.classList.remove('active');
       }
-      else if (TargetInfo.strTargetType === 'a') {
-        $ContextMenuLinkOpen.href = TargetInfo.strTargetValue;
+      else {
+        N.$ContextMenu.dataset.target = TargetInfo.strTargetType;
+
+        // No need to check TargetInfo.strTargetValue, the user can't directly add an empty value and the document content is cleared at input.
+
+        if (TargetInfo.strTargetType === 'a') {
+          $ContextMenuLinkOpen.href = TargetInfo.strTargetValue;
+        }
+        else if (TargetInfo.strTargetType === 'anchor-link') {
+          $ContextMenuAnchorGoTo.dataset.anchor = TargetInfo.strTargetValue;
+        }
+        else if (TargetInfo.strTargetType === 'img') {
+          $ContextMenuImgOpen.href = TargetInfo.strTargetValue;
+        }
+        else if (TargetInfo.strTargetType === 'tab') {
+          N.$Header.classList.add('active');
+        }
+
+        funcPositionElementContext({
+          $Element: N.$ContextMenu,
+          Event: Event
+        });
+
+        N.$ContextMenu.classList.add('active');
+
+        N.LastContextMenuElementInfo = TargetInfo;
+
+        // Toolbar and context menu must not be visible at the same time.
+        N.Functions.Toolbar.funcResetView();
+        N.$Toolbar.classList.remove('active');
       }
-      else if (TargetInfo.strTargetType === 'img') {
-        $ContextMenuImgOpen.href = TargetInfo.strTargetValue;
-      }
-
-      funcPositionElementContext({
-        $Element: $ContextMenu,
-        Event: Event
-      });
-
-      $ContextMenu.classList.add('active');
-
-      N.LastContextMenuElementInfo = TargetInfo;
     } else {
-      $ContextMenu.classList.remove('active');
+      N.$ContextMenu.classList.remove('active');
     }
   });
 
 
 
-  forEach($ContextMenu.querySelectorAll('li.copy'), ($Item) => {
+  N.$ContextMenu.addEventListener('mousedown', (Event) => {
+    Event.preventDefault();
+  });
+
+
+  forEach(N.$ContextMenu.querySelectorAll('li.copy'), ($Item) => {
     $Item.addEventListener('click', () => {
-      $ContextMenu.classList.remove('active');
+      N.$ContextMenu.classList.remove('active');
       N.ElectronFramework.Clipboard.writeText(N.LastContextMenuElementInfo.strTargetValue);
     });
   });
 
-  forEach($ContextMenu.querySelectorAll('li.open'), ($Item) => {
+  forEach(N.$ContextMenu.querySelectorAll('li.open'), ($Item) => {
     $Item.addEventListener('click', () => {
-      $ContextMenu.classList.remove('active');
+      N.$ContextMenu.classList.remove('active');
     });
   });
 
-  $ContextMenu.getElementsByClassName('close-tabs')[0].addEventListener('click', () => {
-    $ContextMenu.classList.remove('active');
+  N.$ContextMenu.getElementsByClassName('close-tabs')[0].addEventListener('click', () => {
+    N.$ContextMenu.classList.remove('active');
     N.Functions.Documents.funcCloseAll();
+  });
+
+  N.$ContextMenu.getElementsByClassName('select-all')[0].addEventListener('click', () => {
+    N.$ContextMenu.classList.remove('active');
+    document.execCommand('selectAll');
+  });
+  N.$ContextMenu.getElementsByClassName('copy-selection')[0].addEventListener('click', () => {
+    N.$ContextMenu.classList.remove('active');
+    document.execCommand('copy');
+  });
+  N.$ContextMenu.getElementsByClassName('paste')[0].addEventListener('click', () => {
+    N.$ContextMenu.classList.remove('active');
+    document.execCommand('paste');
+  });
+
+
+  $ContextMenuAnchorGoTo.addEventListener('click', () => {
+    N.$ContextMenu.classList.remove('active');
+
+    const $Anchor = document.querySelector($ContextMenuAnchorGoTo.dataset.anchor);
+    if ($Anchor) {
+      zenscroll.createScroller(N.DocActive.$ContentContainer).center($Anchor);
+    }
   });
 
 })();