+++ app-dev/js/functions/Dialogs.functions.js
@@ -10,7 +10,7 @@ N.Functions.Dialogs.funcShow = (Parameter) => {
   N.strDialogActiveSlug = Parameter.strDialogSlug;
 
   if (N.strDialogActiveSlug) {
-    N.$DialogActive = document.getElementById(`${N.strDialogActiveSlug}-dialog`);
+    N.$DialogActive = document.getElementById(`${N.strDialogActiveSlug}-dialog_unc2741`);
   } else {
     N.$DialogActive = null;
   }
@@ -45,6 +45,7 @@ N.Functions.Dialogs.funcForceClose = () => {
 N.Functions.Dialogs.funcCloseContext = () => {
   if (
     N.strDialogActiveSlug
+    && N.strDialogActiveSlug !== 'table-content'
     && N.strDialogActiveSlug !== 'doc-info'
     && N.strDialogActiveSlug !== 'unsaved-docs'
     && N.strDialogActiveSlug !== 'new-update'
@@ -53,4 +54,78 @@ N.Functions.Dialogs.funcCloseContext = () => {
   ) {
     N.Functions.Dialogs.funcCloseCurrent();
   }
+};
+
+
+
+N.Functions.Dialogs.funcUpdateTableContent = () => {
+  N.$HeadingsList.innerHTML = '';
+  const arrDocHeadings = N.DocActive.$ContentEditable.querySelectorAll('h1, h2, h3, h4, h5, h6');
+
+  if (arrDocHeadings.length) {
+    N.Functions.Content.funcSetHeadingsIDs({ Document: N.DocActive });
+
+    forEach(arrDocHeadings, ($H) => {
+      const intHeadingLevel = parseInt($H.tagName.slice(-1));
+      let strHeadingLevel = '';
+      switch (intHeadingLevel) {
+        case 1:
+          strHeadingLevel = 'XL';
+          break;
+        case 2:
+          strHeadingLevel = 'L';
+          break;
+        case 3:
+          strHeadingLevel = 'M';
+          break;
+        case 4:
+          strHeadingLevel = 'S';
+          break;
+        case 5:
+          strHeadingLevel = 'XS';
+          break;
+        case 6:
+          strHeadingLevel = 'XXS';
+          break;
+      }
+
+      const $HeadingItem = N.$TableContentHeadingModel.cloneNode(true);
+      $HeadingItem.getElementsByClassName('mdl-list__item-primary-content')[0].textContent = $H.textContent;
+      $HeadingItem.getElementsByClassName('mdl-list__item-secondary-action')[0].textContent = strHeadingLevel;
+      $HeadingItem.dataset.anchor = '#' + $H.id;
+
+
+      $HeadingItem.addEventListener('click', () => {
+        const $Anchor = document.querySelector($HeadingItem.dataset.anchor);
+        if ($Anchor) {
+          zenscroll.createScroller(N.DocActive.$ContentContainer).center($Anchor);
+        }
+      });
+
+      N.$HeadingsList.appendChild($HeadingItem);
+    });
+
+    N.$TableContentDialog.dataset.view = 'headings-list';
+  }
+  else {
+    N.$TableContentDialog.dataset.view = 'no-headings';
+  }
+};
+
+
+
+N.Functions.Dialogs.funcUpdateDocInfo = () => {
+  if (N.DocActive.strDocThemeName) {
+    N.$DocInfoThemeName.textContent = N.DocActive.strDocThemeName;
+  } else {
+    N.$DocInfoThemeName.textContent = '—';
+  }
+
+  Countable.once(N.DocActive.$ContentEditable, (Counter) => {
+    N.$DocInfoParagraphs.textContent = Counter.paragraphs;
+    N.$DocInfoSentences.textContent = Counter.sentences;
+    N.$DocInfoWords.textContent = Counter.words;
+    N.$DocInfoCharacters.textContent = Counter.characters;
+    N.$DocInfoCharactersSpaces.textContent = Counter.all;
+  });
 };