+++ app-dev/js/toolbar.js
@@ -1,18 +1,19 @@
 (() => {
 
-  const $ToolParagraph = document.getElementById('tool-paragraph');
-  const $ToolRemoteImage = document.getElementById('tool-remote-image');
-  const $ToolRemoteVideo = document.getElementById('tool-remote-video');
-  const $ToolRemoteEmbed = document.getElementById('tool-remote-embed');
+  const $ToolParagraph = document.getElementById('tool-paragraph_unc2741');
+  const $ToolRemoteImage = document.getElementById('tool-remote-image_unc2741');
+  const $ToolRemoteVideo = document.getElementById('tool-remote-video_unc2741');
+  const $ToolRemoteEmbed = document.getElementById('tool-remote-embed_unc2741');
 
-  const $ToolBold = document.getElementById('tool-bold');
-  const $ToolItalic = document.getElementById('tool-italic');
-  const $ToolUnderline = document.getElementById('tool-underline');
-  const $ToolStrike = document.getElementById('tool-strike');
-  const $ToolApplyLink = document.getElementById('tool-apply-link');
-  const $ToolSuperscript = document.getElementById('tool-superscript');
-  const $ToolSubscript = document.getElementById('tool-subscript');
-  const $ToolClear = document.getElementById('tool-clear');
+  const $ToolBold = document.getElementById('tool-bold_unc2741');
+  const $ToolItalic = document.getElementById('tool-italic_unc2741');
+  const $ToolUnderline = document.getElementById('tool-underline_unc2741');
+  const $ToolStrike = document.getElementById('tool-strike_unc2741');
+  const $ToolApplyLink = document.getElementById('tool-apply-link_unc2741');
+  const $ToolSuperscript = document.getElementById('tool-superscript_unc2741');
+  const $ToolSubscript = document.getElementById('tool-subscript_unc2741');
+  const $ToolClear = document.getElementById('tool-clear_unc2741');
+  const $ToolEmojis = document.getElementById('tool-emojis_unc2741');
 
 
   const $InputRemoteImage = N.$Toolbar.querySelector('.remote-image-view input');
@@ -31,6 +32,40 @@
   const $CancelApplyLink = N.$Toolbar.querySelector('.apply-link-view .cancel-tool');
   const $ValidateApplyLink = N.$Toolbar.querySelector('.apply-link-view .validate-tool');
 
+  const $InputEmojis = N.$Toolbar.querySelector('.emojis-view input');
+  const $CancelEmojis = N.$Toolbar.querySelector('.emojis-view .cancel-tool');
+  const $SearchEmojis = N.$Toolbar.querySelector('.emojis-view .search-button');
+  const $EmojisList = N.$Toolbar.querySelector('.emojis-view .emojis-list');
+  const Emojis = N.Functions.IO.funcLoadEmojis();
+
+
+
+  const funcToolButtonsCompatibility = (Parameter) => {
+    const $Button = Parameter.$ButtonClicked;
+    let boolCanFormat = true;
+
+    // Some tools incompatibilities with activated lists
+    if (N.$ToolUnorderedList.classList.contains('active') || N.$ToolOrderedList.classList.contains('active')) {
+      if (
+        $Button === $ToolParagraph || $Button === N.$ToolQuote ||
+        $Button === N.$ToolH1 || $Button === N.$ToolH2 || $Button === N.$ToolH3 || $Button === N.$ToolH4 || $Button === N.$ToolH5 || $Button === N.$ToolH6
+      ) {
+        boolCanFormat = false;
+      }
+    }
+
+    // List tools incompatibilities with some activated tools
+    if ($Button === N.$ToolUnorderedList || $Button === N.$ToolOrderedList) {
+      if (
+        N.$ToolQuote.classList.contains('active') || N.$ToolH1.classList.contains('active') || N.$ToolH2.classList.contains('active') || N.$ToolH3.classList.contains('active') || N.$ToolH4.classList.contains('active') || N.$ToolH5.classList.contains('active') || N.$ToolH6.classList.contains('active')
+      ) {
+        boolCanFormat = false;
+      }
+    }
+
+    return boolCanFormat;
+  };
+
 
 
   N.$Toolbar.addEventListener('mousedown', (Event) => {
@@ -43,31 +78,9 @@
   });
 
 
-  forEach(N.$Toolbar.getElementsByTagName('li'), ($Button) => {
+  forEach(N.$Toolbar.querySelectorAll('.tools-list-view li'), ($Button) => {
     $Button.addEventListener('click', () => {
-
-      // Special behaviors handling
-      let boolCanFormat = true;
-
-      // Some tools incompatibilities with activated lists
-      if (N.$ToolUnorderedList.classList.contains('active') || N.$ToolOrderedList.classList.contains('active')) {
-        if (
-          $Button === $ToolParagraph || $Button === N.$ToolQuote ||
-          $Button === N.$ToolH1 || $Button === N.$ToolH2 || $Button === N.$ToolH3 || $Button === N.$ToolH4 || $Button === N.$ToolH5 || $Button === N.$ToolH6
-        ) {
-          boolCanFormat = false;
-        }
-      }
-
-      // List tools incompatibilities with some activated tools
-      if ($Button === N.$ToolUnorderedList || $Button === N.$ToolOrderedList) {
-        if (
-          N.$ToolQuote.classList.contains('active') || N.$ToolH1.classList.contains('active') || N.$ToolH2.classList.contains('active') || N.$ToolH3.classList.contains('active') || N.$ToolH4.classList.contains('active') || N.$ToolH5.classList.contains('active') || N.$ToolH6.classList.contains('active')
-        ) {
-          boolCanFormat = false;
-        }
-      }
-
+      const boolCanFormat = funcToolButtonsCompatibility({ $ButtonClicked: $Button });
 
       if (boolCanFormat) {
         if ($Button.dataset.cmd) {
@@ -79,9 +92,14 @@
           }
         } else if ($Button.dataset.tag) {
           N.DocActive.Editor.format($Button.dataset.tag);
+
+          // if it's a heading tool button
+          if (N.RegExpHeading.test($Button.dataset.tag)) {
+            N.Functions.Dialogs.funcUpdateTableContent();
+          }
         } else {
           N.DocActive.LastSelection = lightrange.saveSelection();
-          N.$Toolbar.dataset.view = $Button.id.replace('tool-', '');
+          N.$Toolbar.dataset.view = $Button.id.replace('tool-', '').replace('_unc2741', '');
           N.Functions.Toolbar.funcAutoPosition();
 
           if ($Button === $ToolRemoteImage) {
@@ -95,6 +113,9 @@
           }
           else if ($Button === $ToolApplyLink) {
             $InputApplyLink.select();
+          }
+          else if ($Button === $ToolEmojis) {
+            $InputEmojis.select();
           }
         }
 
@@ -200,6 +221,21 @@
   });
 
 
+
+  // forEach(Emojis, (Emoji) => {
+  //   console.log(Emoji);
+  // });
+
+  forEach(N.$Toolbar.querySelectorAll('.emojis-view .category-buttons li'), ($Button) => {
+    $Button.addEventListener('click', () => {
+      // https://github.com/zengabor/zenscroll#8-execute-something-when-the-scrolling-is-done
+      // .center(element, duration, offset, onDone)
+      zenscroll.createScroller($EmojisList).center(document.querySelector($Button.dataset.anchor), null, -5);
+    });
+  });
+
+
+
   $InputRemoteImage.addEventListener('keyup', (Event) => {
     // Enter
     if (Event.keyCode === 13) {
@@ -241,6 +277,17 @@
     // Escape
     else if (Event.keyCode === 27) {
       $CancelApplyLink.click();
+    }
+  });
+
+  $InputEmojis.addEventListener('keyup', (Event) => {
+    // Enter
+    if (Event.keyCode === 13) {
+      $SearchEmojis.click();
+    }
+    // Escape
+    else if (Event.keyCode === 27) {
+      $CancelEmojis.click();
     }
   });